[rvstat] Add support for parsing map files from inputs
This allows robustly reconstructing the original memory map, which
means we can accurately report on instructions from linked libraries and
no longer need to ask users to compile with -fno-pie -no-pie.

Additionally add the ability to report analyses on a per-object basis,
or to filter by object (e.g. to show results just for the main binary or
just for libc).